nlp/tools/atomic-pdf-qa-streamlit/requirements.txt
streamlit
atomic-agents
PyPDF2
python-dotenv
instructor
openai
